Move inference to dedicated script and thus cleanup tiler
