+++ deadtrees/deployment/tiler.py
@@ -2,11 +2,9 @@
 import argparse
 import warnings
 from pathlib import Path
-from typing import Dict, Optional, Tuple, Union
+from typing import Optional, Tuple, Union
 
 import xarray
-from torch._C import Value
-from xarray.core.dataarray import DataArray
 
 warnings.filterwarnings("ignore", category=UserWarning)
 
@@ -15,10 +13,6 @@ from dataclasses import dataclass
 
 import numpy as np
 import rioxarray
-import torch
-from deadtrees.data.deadtreedata import val_transform
-from deadtrees.deployment.inference import ONNXInference, PyTorchInference
-from tqdm import tqdm
 
 
 @dataclass
@@ -199,100 +193,3 @@ def unmake_blocks_vectorized(x: np.ndarray, d: int, m: int, n: int) -> np.ndarra
         .transpose(0, 2, 1, 3)
         .reshape(m, n)
     )
-
-
-def main():
-    parser = argparse.ArgumentParser()
-    parser.add_argument("infile", type=Path)
-
-    parser.add_argument(
-        "-m",
-        "--model",
-        dest="model",
-        type=Path,
-        default=Path("checkpoints/bestmodel.ckpt"),
-        help="model artefact",
-    )
-
-    parser.add_argument(
-        "-o",
-        dest="outpath",
-        type=Path,
-        default=Path("."),
-        help="output directory",
-    )
-
-    parser.add_argument(
-        "--all",
-        action="store_true",
-        dest="all",
-        default=False,
-        help="process complete directory",
-    )
-
-    args = parser.parse_args()
-
-    bs = 32
-
-    INFILE = args.infile
-
-    def is_valid_tile(infile):
-        with rioxarray.open_rasterio(infile).sel(band=1) as t:
-            return False if np.isin(t, [0, 255]).all() else True
-
-    # inference = ONNXInference("checkpoints/bestmodel.onnx")
-    inference = PyTorchInference(args.model)
-
-    #
-    # model.to(device)
-
-    if args.all:
-        INFILES = sorted(INFILE.glob("ortho*.tif"))
-    else:
-        INFILES = [INFILE]
-
-    for INFILE in INFILES:
-
-        if not is_valid_tile(INFILE):
-            continue
-
-        tiler = Tiler()
-        tiler.load_file(INFILE)
-
-        batches = tiler.get_batches()
-        batches = np.array_split(batches, math.ceil(len(batches) / bs), axis=0)
-
-        out_batches = []
-
-        for b, batch in enumerate(tqdm(batches, desc=INFILE.name)):
-            batch_tensor = torch.stack(
-                [val_transform(image=i.transpose(1, 2, 0))["image"] for i in batch]
-            )
-
-            # onnx
-            # out_batch = inference.run(batch_tensor.detach().cpu().numpy())
-
-            # pytorch
-            out_batch = (
-                inference.run(batch_tensor.detach().to("cuda"), device="cuda")
-                .cpu()
-                .numpy()
-            )
-
-            out_batches.append(out_batch)
-
-        OUTFILE = args.outpath / INFILE.name
-        OUTFILE_PREVIEW = Path(str(args.outpath) + "_preview") / INFILE.name
-
-        tiler.put_batches(np.concatenate(out_batches, axis=0))
-        tiler.write_file(OUTFILE)
-
-        #
-        from PIL import Image
-
-        image = Image.fromarray(np.uint8(tiler._target.values * 255), "L")
-        image.save(OUTFILE_PREVIEW)
-
-
-if __name__ == "__main__":
-    main()
